+++ vm/src/chips/chips/riscv_cpu/instruction/columns.rs
@@ -6,7 +6,7 @@ use std::{iter::once, mem::size_of, vec::IntoIter};
 pub const NUM_INSTRUCTION_COLS: usize = size_of::<InstructionCols<u8>>();
 
 // RISC-V X0 register
-const REGISTER_X0: u32 = 0;
+const REGISTER_X0: u8 = 0;
 
 #[derive(AlignedBorrow, Clone, Copy, Debug, Default)]
 #[repr(C)]
@@ -30,7 +30,8 @@ pub struct InstructionCols<T> {
 impl<F: Field> InstructionCols<F> {
     pub fn populate(&mut self, instruction: Instruction) {
         self.opcode = instruction.opcode.as_field::<F>();
-        self.op_a = instruction.op_a.into();
+
+        self.op_a = (instruction.op_a as u64).into();
         self.op_b = instruction.op_b.into();
         self.op_c = instruction.op_c.into();
